Unite implementation of `Image.withCGImage`
Using `_CGPixel`.
