+++ Sources/EasyImagy/CoreGraphics.swift.gyb
@@ -320,6 +320,33 @@ extension Image where Pixel: _CGDirectPixel {
 
     }
 
+    public func withCGImage<R>(_ body: (CGImage) throws -> R) rethrows -> R {
+        let length = count * MemoryLayout<Pixel>.size
+
+        var image = self
+        let provider: CGDataProvider = CGDataProvider(data: Data(
+            bytesNoCopy: &image.pixels,
+            count: length,
+            deallocator: .none
+        ) as CFData)!
+
+        let cgImage = CGImage(
+            width: width,
+            height: height,
+            bitsPerComponent: MemoryLayout<Pixel._EZ_PixelDirectChannel>.size * 8,
+            bitsPerPixel: MemoryLayout<Pixel>.size * 8,
+            bytesPerRow: MemoryLayout<Pixel>.size * width,
+            space: Pixel._ez_cgColorSpace,
+            bitmapInfo: Pixel._ez_cgBitmapInfo,
+            provider: provider,
+            decode: nil,
+            shouldInterpolate: false,
+            intent: CGColorRenderingIntent.defaultIntent
+        )!
+
+        return try body(cgImage)
+    }
+
     public mutating func withCGContext(coordinates: CGContextCoordinates = .natural, _ body: (CGContext) throws -> Void) rethrows {
         let width = self.width
         let height = self.height
@@ -422,36 +449,6 @@ extension ImageSlice where Pixel: _CGDirectPixel {
         try body(context)
     }
 }
-% for type in uint_types:
-
-extension Image where Pixel == PremultipliedRGBA<${type}> {
-    public func withCGImage<R>(_ body: (CGImage) throws -> R) rethrows -> R {
-        return try Image.withGeneratedCGImage(
-            image: self,
-            colorSpace: Pixel._ez_cgColorSpace,
-            bitmapInfo: Pixel._ez_cgBitmapInfo,
-            body: body,
-            componentType: ${type}.self
-        )
-    }
-}
-% end
-%
-% for type in uint_types:
-
-extension Image where Pixel == ${type} {
-    public func withCGImage<R>(_ body: (CGImage) throws -> R) rethrows -> R {
-        return try Image.withGeneratedCGImage(
-            image: self,
-            colorSpace: Pixel._ez_cgColorSpace,
-            bitmapInfo: Pixel._ez_cgBitmapInfo,
-            body: body,
-            componentType: ${type}.self
-        )
-    }
-}
-% end
-%
 % for type in uint_types:
 
 extension ImageSlice where Pixel == PremultipliedRGBA<${type}> {
