+++ Sources/EasyImagy/Image.swift
@@ -75,45 +75,3 @@ extension Image {
         return y * width + x
     }
 }
-
-#if canImport(CoreGraphics)
-import Foundation
-import CoreGraphics
-
-extension Image { // Gray or PremultipliedRGBA
-    internal static func withGeneratedCGImage<Component, R>(
-        image: Image<Pixel>,
-        colorSpace: CGColorSpace,
-        bitmapInfo: CGBitmapInfo,
-        body: (CGImage) throws -> R,
-        componentType: Component.Type
-    ) rethrows -> R {
-        var image = image
-        let length = image.count * MemoryLayout<Pixel>.size
-        let width = image.width
-        let height = image.height
-        
-        let provider: CGDataProvider = CGDataProvider(data: Data(
-            bytesNoCopy: &image.pixels,
-            count: length,
-            deallocator: .none
-        ) as CFData)!
-        
-        let cgImage = CGImage(
-            width: width,
-            height: height,
-            bitsPerComponent: MemoryLayout<Component>.size * 8,
-            bitsPerPixel: MemoryLayout<Pixel>.size * 8,
-            bytesPerRow: MemoryLayout<Pixel>.size * width,
-            space: colorSpace,
-            bitmapInfo: bitmapInfo,
-            provider: provider,
-            decode: nil,
-            shouldInterpolate: false,
-            intent: CGColorRenderingIntent.defaultIntent
-        )!
-        
-        return try body(cgImage)
-    }
-}
-#endif
